+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.8.1] - Unreleased
 ### Fixed
 * DLT notebook for CDC merge operations
+* `target_id` and `target_name` for CDC merge options
 * Parsing of data sources in `DataFrameMethod` when NATIVE dataframe API is used
 ### Updated
 * raise `RuntimeError` when wheel file can't be built.